Update client.py
`.next()` would not load the next page of results using a `while True...result = spotify.next(result)`. This has been updated to determine if `limit + offset` is less than `total`, if so, update `result['href']` with the new `offset`.

```python
sp = spotipy.Spotify()

def get_albums(artist, locale=None):
    artist = 'artist:{}'.format(artist)
    result = sp.search(q=artist, type='album', limit=10, market=locale)
    
    while True:
        print len(result['albums']['items'])
        
        result = sp.next(result['albums'])

        if result is None:
            break
```

`market` was also added as an argument in `.search` as the API allows to search in a specific market.